+++ flink-runtime/src/main/java/org/apache/flink/runtime/jobmaster/DefaultSlotPoolServiceSchedulerFactory.java
@@ -26,6 +26,7 @@ import org.apache.flink.configuration.Configuration;
 import org.apache.flink.configuration.JobManagerOptions;
 import org.apache.flink.configuration.RpcOptions;
 import org.apache.flink.configuration.SchedulerExecutionMode;
+import org.apache.flink.configuration.StateRecoveryOptions;
 import org.apache.flink.core.failure.FailureEnricher;
 import org.apache.flink.runtime.blob.BlobWriter;
 import org.apache.flink.runtime.blocklist.BlocklistOperations;
@@ -265,7 +266,7 @@ public final class DefaultSlotPoolServiceSchedulerFactory
     static RequestSlotMatchingStrategy getRequestSlotMatchingStrategy(
             Configuration configuration, JobType jobType) {
         final boolean isLocalRecoveryEnabled =
-                configuration.get(CheckpointingOptions.LOCAL_RECOVERY);
+                configuration.get(StateRecoveryOptions.LOCAL_RECOVERY);
 
         if (isLocalRecoveryEnabled) {
             if (jobType == JobType.STREAMING) {
