+++ flink-tests/src/test/java/org/apache/flink/test/checkpointing/AutoRescalingITCase.java
@@ -38,6 +38,7 @@ import org.apache.flink.configuration.JobManagerOptions;
 import org.apache.flink.configuration.NettyShuffleEnvironmentOptions;
 import org.apache.flink.configuration.StateBackendOptions;
 import org.apache.flink.configuration.StateChangelogOptions;
+import org.apache.flink.configuration.StateRecoveryOptions;
 import org.apache.flink.configuration.WebOptions;
 import org.apache.flink.contrib.streaming.state.RocksDBConfigurableOptions;
 import org.apache.flink.core.testutils.OneShotLatch;
@@ -166,7 +167,7 @@ public class AutoRescalingITCase extends TestLogger {
             config.set(CheckpointingOptions.INCREMENTAL_CHECKPOINTS, true);
             // todo: local rescaling is not supported by changelog.
             config.set(StateChangelogOptions.ENABLE_STATE_CHANGE_LOG, false);
-            config.set(CheckpointingOptions.LOCAL_RECOVERY, true);
+            config.set(StateRecoveryOptions.LOCAL_RECOVERY, true);
             config.set(
                     CheckpointingOptions.CHECKPOINTS_DIRECTORY, checkpointDir.toURI().toString());
             config.set(CheckpointingOptions.SAVEPOINT_DIRECTORY, savepointDir.toURI().toString());
